backend/requirements.txt
django
djangorestframework
mysqlclient
django-cors-headers
djangorestframework-simplejwt
